@@ -24,51 +27,64 @@
   },
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=6.14.3"
+    "node": ">=14.16"
   },
+  "type": "module",
   "devDependencies": {
-    "@types/body-parser": "^1.16.0",
-    "@types/express": "^4.0.34",
-    "@types/jasmine": "^3.3.12",
-    "@types/js-yaml": "^4.0.0",
-    "@types/lodash": "^4.14.126",
-    "@types/node": "^14.14.23",
-    "@types/q": "^1.0.0",
-    "@types/validator": "^13.1.0",
-    "@types/verror": "^1.9.0",
-    "ansi-colors": "^4.1.1",
-    "body-parser": "^1.18.3",
-    "conventional-changelog": "^3.1.23",
-    "conventional-changelog-lint": "^2.1.1",
-    "express": "^4.16.3",
+    "@babel/core": "^7.21.8",
+    "@babel/preset-env": "^7.21.5",
+    "@babel/preset-typescript": "^7.21.5",
+    "@babel/register": "^7.21.0",
+    "@commitlint/cli": "^17.6.1",
+    "@commitlint/config-conventional": "^17.6.1",
+    "@rollup/plugin-babel": "^6.0.3",
+    "@rollup/plugin-commonjs": "^24.1.0",
+    "@rollup/plugin-json": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.2",
+    "@types/body-parser": "^1.19.2",
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "^4.3.1",
+    "@types/js-yaml": "^4.0.5",
+    "@types/lodash": "^4.14.194",
+    "@types/node": "^18.16.3",
+    "@types/q": "^1.5.5",
+    "@types/validator": "^13.7.15",
+    "@types/verror": "^1.10.6",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
+    "ansi-colors": "^4.1.3",
+    "body-parser": "^1.20.2",
+    "conventional-changelog": "^3.1.25",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-jest": "^27.2.1",
+    "express": "^4.18.2",
     "gulp": "^4.0.2",
-    "gulp-bump": "^3.1.1",
-    "gulp-clean": "^0.4.0",
-    "gulp-cli": "^2.0.1",
-    "gulp-conventional-changelog": "^2.0.19",
-    "gulp-filter": "^6.0.0",
-    "gulp-git": "^2.7.0",
-    "gulp-jasmine": "^4.0.0",
-    "gulp-tslint": "^8.1.3",
-    "gulp-typescript": "^5.0.1",
-    "gulp-util": "^3.0.7",
-    "minimist": "^1.2.0",
+    "gulp-bump": "^3.2.0",
+    "gulp-cli": "^2.3.0",
+    "gulp-conventional-changelog": "^2.0.35",
+    "gulp-git": "^2.10.1",
+    "husky": "^8.0.3",
+    "jasmine": "^4.6.0",
+    "minimist": "^1.2.8",
+    "prettier": "^2.8.8",
+    "rollup": "^3.21.4",
     "tslint": "^6.1.3",
-    "typescript": "^4.1.3"
+    "typescript": "^5.0.4"
   },
   "dependencies": {
+    "@apidevtools/swagger-parser": "^10.1.0",
     "ajv": "^6.12.6",
-    "axios": "^0.27.2",
-    "commander": "^7.0.0",
-    "decimal.js": "^10.2.0",
-    "js-yaml": "^4.0.0",
-    "jsonpointer": "^5.0.0",
-    "lodash": "^4.17.10",
-    "openapi-types": "^7.0.1",
-    "swagger-parser": "^10.0.2",
-    "uuidjs": "^4.0.3",
-    "validator": "^13.1.17",
-    "verror": "^1.8.1"
+    "axios": "^1.4.0",
+    "commander": "^10.0.1",
+    "decimal.js": "^10.4.3",
+    "js-yaml": "^4.1.0",
+    "jsonpointer": "^5.0.1",
+    "lodash": "^4.17.21",
+    "openapi-types": "^12.1.0",
+    "uuidjs": "^5.0.1",
+    "validator": "^13.9.0",
+    "verror": "^1.10.1"
   },
   "types": "lib/api-types.d.ts"
 }
