@@ -1,7 +1,6 @@
-import * as os from 'os';
+import os from 'os';
 
-// tslint:disable:no-var-requires
-const packageJson = require('../../../package.json');
+import packageJson from "../../../package.json";
 
 export class Metadata {
     public getHostname(): string {
